refactor(migrate): split upsampler position
Move citation position analysis into a private upsampler submodule.

Keep the public upsampler API stable for csl26-rfct.
